preparing the xcode file
