+++ TomatoMethodApp/Resource/Assets.xcassets/Contents.json
