+++ TomatoMethodApp/Source/App/AppDelegate.swift
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.swift
+//  TomatoMethodApp
+//
+//  Created by Kristina Korotkova on 23/12/22.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+
+        return true
+    }
+}
