+++ TomatoMethodApp/Source/App/SceneDelegate.swift
@@ -1,0 +1,45 @@
+//
+//  SceneDelegate.swift
+//  TomatoMethodApp
+//
+//  Created by Kristina Korotkova on 23/12/22.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        let window = UIWindow(windowScene: windowScene)
+        window.rootViewController = ViewController()
+        window.makeKeyAndVisible()
+        self.window = window
+    }
+
+    func sceneDidDisconnect(_ scene: UIScene) {
+
+    }
+
+    func sceneDidBecomeActive(_ scene: UIScene) {
+
+    }
+
+    func sceneWillResignActive(_ scene: UIScene) {
+
+    }
+
+    func sceneWillEnterForeground(_ scene: UIScene) {
+
+    }
+
+    func sceneDidEnterBackground(_ scene: UIScene) {
+
+    }
+
+
+}
+
